Once again reverted the api, spelling, useless destructor
For now I'm not going to implement the API on classes until they need it as this is causing a lot of problems and isn't fully necessary yet.

Fix a spelling mistake in window.

Got rid of the virtual destructor for vector2 as it is a data class and shouldn't ever get inheritance, if it ends up being needed I'll add it then.