Event and minor comment
Added event and EventDispatcher header files as well as outlined and implemented the event file to the best of my current knowledge. Started outlining the dispatcher.

Added a comment to Vector2 to remind me to move it to a header and template based class as that is better for efficiency.